fix mte instructions.

+++ riscv/insns/dlxfer_d.h
@@ -2,7 +2,7 @@ auto proxy = p->get_proxy();
 if (proxy == nullptr) return -1;
 
 //TCP_AUNIT();
-auto dst = *((unsigned long long *)MMU.get_phy_addr(RD));
+auto dst = *((unsigned int *)MMU.get_phy_addr(RD));
 auto src = RS1;
 auto len = RS2;
