+++ riscv/insns/dlxfer_u.h
@@ -3,7 +3,7 @@ if (proxy == nullptr) return -1;
 
 //TCP_AUNIT();
 auto dst = RD;
-auto src = *((unsigned long long *)MMU.get_phy_addr(RS1));
+auto src = *((unsigned int *)MMU.get_phy_addr(RS1));
 auto len = RS2;
 
 proxy->dmaXfer(src, dst, len, Transport::AbstractProxy::DDR2LLB);
